blip-model/requirements.txt
transformers==4.40.1
torch
flask
Pillow
requests